refactoring the AuthControler
