+++ ecommers/src/main/resources/application.properties
@@ -2,6 +2,7 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/ecommerce
 spring.datasource.dbname=ecommerce
 spring.datasource.username=root
+spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.show-sql=true
